requirements.txt
streamlit
spacy
grpcio-status
google-cloud-secret-manager
transformers
scikit-learn
sentence-transformers
pdfplumber
docx2txt
pandas
numpy